Updated the scan button position in the GraphattributesWidget
